maps are now updating + model updates

+++ server/hitchbot-secure-api/hitchbot-secure-api/Controllers/ReturnObjects/Location.cs
@@ -14,11 +14,11 @@ namespace hitchbot_secure_api.Controllers
     {
         public class ReturnLocation : GenericHitchBot
         {
-            public decimal Latitude { get; set; }
-            public decimal Longitude { get; set; }
-            public decimal? Altitude { get; set; }
-            public decimal? Accuracy { get; set; }
-            public decimal? Velocity { get; set; }
+            public double Latitude { get; set; }
+            public double Longitude { get; set; }
+            public double? Altitude { get; set; }
+            public double? Accuracy { get; set; }
+            public double? Velocity { get; set; }
 
             public int? LocationProviderEnum { get; set; }
